+++ tests/NanoToTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Niush\LaravelNanoTo\Tests\Feature;
+namespace Niush\NanoTo\Tests\Feature;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7\Response;
-use Niush\LaravelNanoTo\LaravelNanoToFacade as LaravelNanoTo;
-use Niush\LaravelNanoTo\Tests\TestCase;
+use Niush\NanoTo\NanoToFacade as NanoTo;
+use Niush\NanoTo\Tests\TestCase;
 
-class LaravelNanoToTest extends TestCase
+class NanoToTest extends TestCase
 {
     protected $app;
     protected $testing_checkout_url = "https://example.com/test_id";
@@ -42,7 +42,7 @@ class LaravelNanoToTest extends TestCase
     {
         $app['config']->set('app.env', 'local');
         $app['config']->set('app.debug', true);
-        $app['config']->set('laravel-nano-to.accounts', [
+        $app['config']->set('nano-to.accounts', [
             'nano' => [
                 'nano_378shkx4k3wd5gxmj3xnjwuxtaf9xrehyz7ugakpiemh8arxq8w9a9xniush',
             ],
@@ -65,11 +65,11 @@ class LaravelNanoToTest extends TestCase
     /** @test */
     public function fails_if_accounts_config_is_empty()
     {
-        $this->app['config']->set('laravel-nano-to.accounts', [
+        $this->app['config']->set('nano-to.accounts', [
             'nano' => [],
         ]);
         try {
-            $response = LaravelNanoTo::amount(9.99)->create('unique_id')->send();
+            $response = NanoTo::amount(9.99)->create('unique_id')->send();
         } catch (\Error$e) {
             $this->assertStringContainsString($e->getMessage(), 'Receiver Account was not available.');
         }
@@ -78,9 +78,9 @@ class LaravelNanoToTest extends TestCase
     /** @test */
     public function redirects_to_checkout_and_create_functions_callback_is_also_equal()
     {
-        $response = LaravelNanoTo::amount(9.99)->create('unique_id', function ($checkout_url) {
+        $response = NanoTo::amount(9.99)->create('unique_id', function ($checkout_url) {
             if ($this->use_real_api) {
-                $this->assertStringStartsWith($this->app['config']->get('laravel-nano-to.base_url'), $checkout_url);
+                $this->assertStringStartsWith($this->app['config']->get('nano-to.base_url'), $checkout_url);
             } else {
                 $this->assertEquals($this->testing_checkout_url, $checkout_url);
             }
@@ -88,7 +88,7 @@ class LaravelNanoToTest extends TestCase
 
         $this->assertEquals(302, $response->status());
         if ($this->use_real_api) {
-            $this->assertStringStartsWith($this->app['config']->get('laravel-nano-to.base_url'), $response->headers->get('location'));
+            $this->assertStringStartsWith($this->app['config']->get('nano-to.base_url'), $response->headers->get('location'));
         } else {
             $this->assertEquals($this->testing_checkout_url, $response->headers->get('location'));
         }
@@ -100,17 +100,17 @@ class LaravelNanoToTest extends TestCase
         $expected_body = [
             "title" => "Payment for Subscription",
             "description" => "<i>Also accepts HTML</i>",
-            "success_url" => route(config('laravel-nano-to.success_url'), 'unique_id'),
-            "cancel_url" => route(config('laravel-nano-to.cancel_url'), 'unique_id'),
-            "webhook_url" => route(config('laravel-nano-to.webhook_url'), 'unique_id'),
+            "success_url" => route(config('nano-to.success_url'), 'unique_id'),
+            "cancel_url" => route(config('nano-to.cancel_url'), 'unique_id'),
+            "webhook_url" => route(config('nano-to.webhook_url'), 'unique_id'),
             "webhook_secret" => "",
-            "background" => config('laravel-nano-to.background'),
-            "color" => config('laravel-nano-to.color'),
+            "background" => config('nano-to.background'),
+            "color" => config('nano-to.color'),
             "raw" => false,
             "business" => [
-                "name" => config('laravel-nano-to.business.name'),
-                "logo" => config('laravel-nano-to.business.logo'),
-                "favicon" => config('laravel-nano-to.business.favicon'),
+                "name" => config('nano-to.business.name'),
+                "logo" => config('nano-to.business.logo'),
+                "favicon" => config('nano-to.business.favicon'),
             ],
             "plans" => [
                 ["name" => "Coffee", "price" => "10"],
@@ -118,7 +118,7 @@ class LaravelNanoToTest extends TestCase
             ],
         ];
 
-        $response = LaravelNanoTo::info("Payment for Subscription", "<i>Also accepts HTML</i>")
+        $response = NanoTo::info("Payment for Subscription", "<i>Also accepts HTML</i>")
             ->suggest([
                 ["name" => "Coffee", "price" => "10"],
                 ["name" => "Meal", "price" => "50"],
@@ -132,25 +132,25 @@ class LaravelNanoToTest extends TestCase
     public function webhook_secret_env_appends_if_exists()
     {
         $expected_body = [
-            "title" => config('laravel-nano-to.title'),
-            "description" => config('laravel-nano-to.description'),
-            "success_url" => route(config('laravel-nano-to.success_url'), 'unique_id'),
-            "cancel_url" => route(config('laravel-nano-to.cancel_url'), 'unique_id'),
-            "webhook_url" => route(config('laravel-nano-to.webhook_url'), 'unique_id'),
+            "title" => config('nano-to.title'),
+            "description" => config('nano-to.description'),
+            "success_url" => route(config('nano-to.success_url'), 'unique_id'),
+            "cancel_url" => route(config('nano-to.cancel_url'), 'unique_id'),
+            "webhook_url" => route(config('nano-to.webhook_url'), 'unique_id'),
             "webhook_secret" => "123456",
-            "background" => config('laravel-nano-to.background'),
-            "color" => config('laravel-nano-to.color'),
+            "background" => config('nano-to.background'),
+            "color" => config('nano-to.color'),
             "raw" => false,
             "business" => [
-                "name" => config('laravel-nano-to.business.name'),
-                "logo" => config('laravel-nano-to.business.logo'),
-                "favicon" => config('laravel-nano-to.business.favicon'),
+                "name" => config('nano-to.business.name'),
+                "logo" => config('nano-to.business.logo'),
+                "favicon" => config('nano-to.business.favicon'),
             ],
         ];
 
-        $this->app['config']->set('laravel-nano-to.webhook_secret', '123456');
+        $this->app['config']->set('nano-to.webhook_secret', '123456');
 
-        $response = LaravelNanoTo::create('unique_id', function ($checkout_url, $body) use ($expected_body) {
+        $response = NanoTo::create('unique_id', function ($checkout_url, $body) use ($expected_body) {
             $this->assertEquals($expected_body, $body);
         });
     }
@@ -158,10 +158,10 @@ class LaravelNanoToTest extends TestCase
     /** @test */
     public function can_use_custom_webhook_secret()
     {
-        $this->app['config']->set('laravel-nano-to.webhook_secret', '123456');
+        $this->app['config']->set('nano-to.webhook_secret', '123456');
 
-        $response = LaravelNanoTo::secret(
-            config("laravel-nano-to.webhook_secret") . "-custom"
+        $response = NanoTo::secret(
+            config("nano-to.webhook_secret") . "-custom"
         )->create('unique_id', function ($checkout_url, $body) {
             $this->assertEquals($body["webhook_secret"], '123456-custom');
         });
@@ -175,13 +175,13 @@ class LaravelNanoToTest extends TestCase
             "logo" => "https://example.com/logo.png",
             "favicon" => "https://example.com/logo.ico",
         ];
-        $this->app['config']->set('laravel-nano-to.business', $business);
+        $this->app['config']->set('nano-to.business', $business);
 
-        $response = LaravelNanoTo::create('unique_id', function ($checkout_url, $body) use ($business) {
+        $response = NanoTo::create('unique_id', function ($checkout_url, $body) use ($business) {
             $this->assertEquals($body["business"], $business);
         });
 
-        $response = LaravelNanoTo::business([
+        $response = NanoTo::business([
             "name" => "My Custom Company",
         ])->create('unique_id', function ($checkout_url, $body) use ($business) {
             $this->assertEquals($body["business"]["name"], "My Custom Company");
@@ -194,10 +194,10 @@ class LaravelNanoToTest extends TestCase
         $background = "#000000,#FFFFFF";
         $color = "#FFFFFF,#000000";
 
-        $this->app['config']->set('laravel-nano-to.background', $background);
-        $this->app['config']->set('laravel-nano-to.color', $color);
+        $this->app['config']->set('nano-to.background', $background);
+        $this->app['config']->set('nano-to.color', $color);
 
-        $response = LaravelNanoTo::create('unique_id', function ($checkout_url, $body) use ($background, $color) {
+        $response = NanoTo::create('unique_id', function ($checkout_url, $body) use ($background, $color) {
             $this->assertEquals($body["background"], $background);
             $this->assertEquals($body["color"], $color);
         });
@@ -205,7 +205,7 @@ class LaravelNanoToTest extends TestCase
         $background = "#111111,#EEEEEE";
         $color = "#EEEEEE,#111111";
 
-        $response = LaravelNanoTo::background($background)->color($color)
+        $response = NanoTo::background($background)->color($color)
             ->create('unique_id', function ($checkout_url, $body) use ($background, $color) {
                 $this->assertEquals($body["background"], $background);
                 $this->assertEquals($body["color"], $color);
@@ -219,7 +219,7 @@ class LaravelNanoToTest extends TestCase
             "payment_type" => "monthly",
         ];
 
-        $response = LaravelNanoTo::metadata($metadata)->create('unique_id', function ($checkout_url, $body) use ($metadata) {
+        $response = NanoTo::metadata($metadata)->create('unique_id', function ($checkout_url, $body) use ($metadata) {
             $this->assertEquals($body["metadata"], $metadata);
         });
     }
@@ -227,7 +227,7 @@ class LaravelNanoToTest extends TestCase
     /** @test */
     public function can_ask_to_use_raw_amount_in_qr()
     {
-        $response = LaravelNanoTo::asRaw()->amount(9.99)->create("unique_id", function ($checkout_url, $body) {
+        $response = NanoTo::asRaw()->amount(9.99)->create("unique_id", function ($checkout_url, $body) {
             $this->assertEquals($body["raw"], true);
         })->send();
     }
@@ -236,7 +236,7 @@ class LaravelNanoToTest extends TestCase
     public function can_add_custom_image_to_checkout_url()
     {
         $image_url = "https://dummyimage.com/300";
-        $response = LaravelNanoTo::withImage($image_url)->amount(9.99)->create("unique_id", function ($checkout_url, $body) use ($image_url) {
+        $response = NanoTo::withImage($image_url)->amount(9.99)->create("unique_id", function ($checkout_url, $body) use ($image_url) {
             $this->assertEquals($body["image"], $image_url);
         })->send();
     }